Added files via upload
Project's APK: https://drive.google.com/file/d/0B9nUhNGV9WK-bHdaeXZlZXpSNXM/view
Project's keynote: https://goo.gl/vi2KH1​

According to SimSpace challenge requirements, we've created a game for Android smartphone and Google Cardboard. Marooned is a strategic game, featuring not only scientific and technical challenges the astronauts face, but also the concept of team work in extreme conditions, and the importance of caring about your teammates. 
According to our scenario, the team of the astronauts arrives to Mars to create a GPS-like navigation system using CubeSat's and ground transmitters. The Team Leader's mission is to coordinate the astronauts' actions, and the other team members' are supposed to accomplish their individual missions within the deadline.
Marooned can be a real conditions simulator for those who prepare for a journey to Mars, or an inspirational challenge for the dreamers, who want to understand the concept of human mission on Mars and it's importance in te